// reservation-service/src/Interfaces/user-interface.ts
import { Observable } from 'rxjs';

export interface UserServiceInterface {
  getUserById(data: GetUserByIdRequest): Observable<GetUserByIdResponse>;
  getUserByEmail(data: GetUserByEmail): Observable<GetUserByEmailResponse>;
  createUser(data: CreateUser): Observable<CreateUserResponse>;
}

export interface CreateUser {
  name: string;
  email: string;
  password: string;
}

export interface User {
  name: string;
  email: string;
  password: string;
}

export interface CreateUserResponse {
  id: number;
  name?: string;
  email?: string;
  error?: boolean;
  message?: string;
}

export interface GetUserByIdRequest {
  id: number;
}

export interface GetUserByIdResponse {
  id: number;
  name: string;
  email: string;
  error?: boolean;
  message?: string;
}

export interface GetUserByEmail {
  email: string;
}

export interface GetUserByEmailResponse {
  id: number;
  name: string;
  email: string;
  password: string;
}